feat: update comment, booking logic

+++ src/main/java/ru/practicum/shareit/item/service/impl/ItemServiceImpl.java
@@ -3,7 +3,12 @@ package ru.practicum.shareit.item.service.impl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.booking.mapper.BookingMapper;
+import ru.practicum.shareit.booking.model.Booking;
+import ru.practicum.shareit.booking.model.Status;
+import ru.practicum.shareit.booking.repository.BookingRepository;
 import ru.practicum.shareit.common.exception.IncorrectParameterException;
+import ru.practicum.shareit.common.exception.NotAvailableException;
 import ru.practicum.shareit.common.exception.NotFoundException;
 import ru.practicum.shareit.common.util.Constants;
 import ru.practicum.shareit.item.dto.CommentDto;
@@ -19,8 +24,8 @@ import ru.practicum.shareit.item.service.ItemService;
 import ru.practicum.shareit.user.model.User;
 import ru.practicum.shareit.user.repository.UserRepository;
 
-import java.util.Collection;
-import java.util.Collections;
+import java.time.LocalDateTime;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,24 +34,74 @@ import java.util.stream.Collectors;
 public class ItemServiceImpl implements ItemService {
     private final ItemMapper itemMapper;
     private final CommentMapper commentMapper;
+    private final BookingMapper bookingMapper;
     private final ItemRepository itemRepository;
     private final UserRepository userRepository;
     private final CommentRepository commentRepository;
+    private final BookingRepository bookingRepository;
 
     @Override
     public Collection<ItemResponseDto> getAllUserItems(long userId) {
         checkIfUserExists(userId);
-        return itemRepository.findAllByOwnerId(userId).stream()
-                .map(itemMapper::convertToResponseDto)
+
+        Collection<Item> items = itemRepository.findAllByOwnerIdOrderByIdAsc(userId);
+        Collection<Booking> bookings = bookingRepository.findAllByItemOwnerIdAndItemIn(userId, items);
+        Map<Long, List<Comment>> commentMap = commentRepository.findAllByItemInEager(items).stream()
+                .collect(Collectors.groupingBy(it -> it.getItem().getId()));
+
+        return items.stream()
+                .map(item -> {
+                    Booking lastBooking = bookings.stream()
+                            .filter(it -> it.getItem().getId().equals(item.getId())
+                                    && it.getStatus().equals(Status.APPROVED)
+                                    && it.getStart().isBefore(LocalDateTime.now()))
+                            .max(Comparator.comparing(Booking::getStart))
+                            .orElse(null);
+                    Booking nextBooking = bookings.stream()
+                            .filter(it -> it.getItem().getId().equals(item.getId())
+                                    && it.getStatus().equals(Status.APPROVED)
+                                    && it.getStart().isAfter(LocalDateTime.now()))
+                            .min(Comparator.comparing(Booking::getStart))
+                            .orElse(null);
+                    Collection<Comment> comments = commentMap.getOrDefault(item.getId(), new ArrayList<>());
+
+                    return itemMapper.convertToResponseDto(item).toBuilder()
+                            .lastBooking(bookingMapper.convertToResponseShortDto(lastBooking))
+                            .nextBooking(bookingMapper.convertToResponseShortDto(nextBooking))
+                            .comments(comments.stream()
+                                    .map(commentMapper::convertToDto)
+                                    .collect(Collectors.toList())).build();
+                })
                 .collect(Collectors.toList());
     }
 
     @Override
-    public ItemResponseDto getItemById(long itemId) {
+    public ItemResponseDto getItemById(long userId, long itemId) {
         Item item = itemRepository.findById(itemId)
                 .orElseThrow(() -> new NotFoundException(String.format("Вещь по id = %s не существует", itemId)));
-        log.info("Запрос вещи по id = {} - {}", itemId, item);
-        return itemMapper.convertToResponseDto(item);
+        Collection<Booking> bookings = item.getOwner().getId() == userId
+                ? bookingRepository.findAllByItemOwnerId(userId) : new ArrayList<>();
+        Map<Long, List<Comment>> commentMap = commentRepository.findAllByItemIdEager(itemId).stream()
+                .collect(Collectors.groupingBy(it -> it.getItem().getId()));
+
+        Booking lastBooking = bookings.stream()
+                .filter(it -> it.getStatus().equals(Status.APPROVED)
+                        && it.getStart().isBefore(LocalDateTime.now()))
+                .max(Comparator.comparing(Booking::getStart))
+                .orElse(null);
+        Booking nextBooking = bookings.stream()
+                .filter(it -> it.getStatus().equals(Status.APPROVED)
+                        && it.getStart().isAfter(LocalDateTime.now()))
+                .min(Comparator.comparing(Booking::getStart))
+                .orElse(null);
+        Collection<Comment> comments = commentMap.getOrDefault(item.getId(), new ArrayList<>());
+
+        return itemMapper.convertToResponseDto(item).toBuilder()
+                .lastBooking(bookingMapper.convertToResponseShortDto(lastBooking))
+                .nextBooking(bookingMapper.convertToResponseShortDto(nextBooking))
+                .comments(comments.stream()
+                        .map(commentMapper::convertToDto)
+                        .collect(Collectors.toList())).build();
     }
 
     @Override
@@ -56,6 +111,8 @@ public class ItemServiceImpl implements ItemService {
                         String.format("Пользователя по id = %s не существует", userId)));
         Item item = itemMapper.convertRequestDtoToEntity(itemDto);
         item.setOwner(user);
+
+        log.info("Добавлена вещь - {}", item);
         return itemMapper.convertToResponseDto(itemRepository.save(item));
     }
 
@@ -72,6 +129,8 @@ public class ItemServiceImpl implements ItemService {
             throw new NotFoundException(String.format("Пользователь по id = %s не является владельцом вещи - %s",
                     userId, updatingItem));
         }
+
+        log.info("Вещь по id = {} обновлена - {}", itemId, updatingItem);
         return itemMapper.convertToResponseDto(itemRepository.save(updatingItem));
     }
 
@@ -96,8 +155,21 @@ public class ItemServiceImpl implements ItemService {
                         String.format("Пользователя по id = %s не существует", userId)));
         Item item = itemRepository.findById(itemId)
                 .orElseThrow(() -> new NotFoundException(String.format("Вещь по id = %s не существует", itemId)));
+        if (!bookingRepository.existsByBookerIdAndItemIdAndStatusAndEndIsBeforeOrderByEndDesc(userId, itemId,
+                Status.APPROVED, LocalDateTime.now())) {
+            log.info("Добавление комментария к вещи по id = {} для пользователя по id = {} недоступно, так как " +
+                    "отсутствует бронь", itemId, userId);
+            throw new NotAvailableException(String.format(
+                    "Добавление комментария к вещи по id = %s для пользователя по id = %s недоступно, так как " +
+                            "отсутствует бронь", itemId, userId));
+        }
+        ;
+
         Comment comment = commentMapper.convertToEntity(commentDto);
+        comment.setAuthor(user);
         comment.setItem(item);
+
+        log.info("Пользователем по id = {} для вещи по id = {} добавлен комментарий - {}", userId, itemId, comment);
         return commentMapper.convertToDto(commentRepository.save(comment));
     }
